refactor: dir rename
