+++ cmd/xkcd/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"courses/internal/adapter/handler"
 	"courses/internal/core"
+	"courses/internal/core/catalog"
 	"courses/internal/core/filler"
-	"courses/internal/core/find"
 	"courses/internal/core/xkcd"
 	"courses/internal/database"
 	"encoding/json"
