+++ internal/core/catalog/find.go
