+++ internal/core/catalog/find_test.go
